Add calc for charge and 2nd can frame formatting

+++ BeagleBone_black/src/data/data.hpp
@@ -132,9 +132,9 @@ struct Battery {
   uint16_t  voltage;      // in 0.1V
   int16_t   current;      // in 0.1A (can be negative)
   uint8_t   charge;       // in % (from 0 to 100)
-  int8_t    temperature;  // in C
-  uint8_t   low_voltage_cell;  // in mV
-  uint8_t   high_voltage_cell;  // in mV
+  int8_t    temperature;  // max temp in C
+  uint16_t  low_voltage_cell;    // in mV
+  uint16_t  high_voltage_cell;   // in mV
 };
 
 struct Batteries : public Module {
